+++ app/jobs/calc_item_level_for_characters_job.rb
@@ -1,0 +1,8 @@
+# Calculate item level for new equipment
+class CalcItemLevelForCharactersJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    CalcItemLevelForCharacters.new.call
+  end
+end