+++ db/migrate/20200128120600_add_item_level_to_characters.rb
@@ -1,0 +1,5 @@
+class AddItemLevelToCharacters < ActiveRecord::Migration[5.2]
+  def change
+    add_column :characters, :item_level, :integer, null: false, default: 0
+  end
+end