+++ db/migrate/20200128121224_add_boolean_to_check_equipment.rb
@@ -1,0 +1,5 @@
+class AddBooleanToCheckEquipment < ActiveRecord::Migration[5.2]
+  def change
+    add_column :characters, :item_level_calculated, :boolean, null: false, default: false
+  end
+end