+++ frontend/components_react/line_up/line_up.jsx
@@ -248,6 +248,7 @@ export default class LineUp extends React.Component {
             </div>
           </td>
           <td>{subscribe.character.level}</td>
+          <td>{subscribe.character.item_level > 0 && subscribe.character.item_level}</td>
           <td>{subscribe.character.guild_name}</td>
           <td>{strings[subscribe.status]}</td>
           <td className="comment_box">{this._renderComment(subscribe)}</td>
@@ -280,6 +281,7 @@ export default class LineUp extends React.Component {
             </div>
           </td>
           <td>{character.level}</td>
+          <td>{character.item_level > 0 && character.item_level}</td>
           <td>{character.guild_name}</td>
         </tr>
       )
@@ -300,7 +302,7 @@ export default class LineUp extends React.Component {
   _renderEmptyLine() {
     return (
       <tr className='empty_line'>
-        <td colSpan={7}></td>
+        <td colSpan={8}></td>
       </tr>
     )
   }
@@ -590,6 +592,7 @@ export default class LineUp extends React.Component {
                 <th>{strings.name}</th>
                 <th>{strings.role}</th>
                 <th>{strings.level}</th>
+                <th>{strings.itemLevel}</th>
                 <th>{strings.guild}</th>
                 <th>{strings.status}</th>
                 <th>{strings.comment}</th>
@@ -620,6 +623,7 @@ export default class LineUp extends React.Component {
                   <th>{strings.name}</th>
                   <th>{strings.role}</th>
                   <th>{strings.level}</th>
+                  <th>{strings.itemLevel}</th>
                   <th>{strings.guild}</th>
                 </tr>
               </thead>